splits mirror and config functionality in 2 files fixes mirror check parallelizing style fix
