+++ cacheman.go
@@ -24,7 +24,7 @@ func HandleReq(w http.ResponseWriter, r *http.Request) {
 	//TODO: review public and private functions
 	RequestedLocalPath := Cfg.CacheDir + "/" + r.URL.Path[1:] //add cachedir to path, to not check in /
 	RequestedPath := r.URL.Path[1:]
-	fmt.Printf("[SERVER %s]File requested: %s\n", NowStr, RequestedPath)
+	fmt.Printf("[SERVER %s] File requested: %s\n", NowStr, RequestedPath)
 
 	RemoteRequired := true
